backend/requirements.txt
# ══════════════════════════════════════════════════════
#  AdGuard requirements.txt
#  pip install -r requirements.txt
# ══════════════════════════════════════════════════════

# ── FastAPI 서비스 ─────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0

# ── Azure OpenAI ───────────────────────────────────────
openai>=1.12.0

# ── Azure AI Search ────────────────────────────────────
azure-search-documents>=11.4.0
azure-core>=1.30.0

# ── Azure Document Intelligence (OCR) ─────────────────
azure-ai-documentintelligence>=1.0.0b1

# ── Azure Storage (Blob + Table) ──────────────────────
azure-storage-blob>=12.19.0
azure-data-tables>=12.5.0

# ── 데이터 처리 ────────────────────────────────────────
pyyaml>=6.0.1
numpy>=1.26.0
tiktoken>=0.6.0

# ── PDF 리포트 생성 ────────────────────────────────────
reportlab>=4.0.0

# ── (선택) 한국어 토크나이저·청킹 스크립트용 ──────────
# kiwipiepy>=0.16.0
# pdfplumber>=0.10.0
# openpyxl>=3.1.2
